refactor(D-170 A): decompose chat_with_agent God-method into staged preempt mixins
1,926→440 lines in chat_turn.py. The ~1,030-line async-generator method is now a
thin orchestrator walking 13 sub-generator stages over a shared TurnContext
(explicit turn_complete contract — order preserved, blocks moved intact):
- turn_preempts_deterministic: policy gate (D-144/D-145/D-158) + greeting +
  question-only + computational (D-143)
- turn_preempts_concept: escalation matrix (D-138) + definitional (D-132) +
  concept example (D-136) + socratic evaluation (D-130) + indexed retrieval (D-101)
- turn_preempts_cognitive: deterministic escape hatch + cognitive engines
  (D-124..D-135/D-155/D-160)
- turn_preempts_delivery: calculated UI MODE_A/B (V38.0) + explanation-with-context
  (D-052/D-103)
- turn_fallback: guarded local fallback chain (REQUIRE_ORCHESTRATOR=0 only)

Companion fix: the PolicyObservation rebinding shadowed the unified-observability
handle, silently killing every obs.end_span after it — renamed to policy_obs,
restoring per-turn telemetry.

Manifest: 6 new files appended to TUTOR_SOURCE_FILES BEFORE chat_turn.py
(execution order = textual order for cross-file ordering assertions) + rule:
no parentheses inside the tuple comments (the textual gate reader stops at the
first closing paren). ISS-110 ordering test realigned to the dispatch contract
(stronger: tuple order + real call sites inside each stage).
